+++ sdk/core/azure-core/src/io/body_stream.cpp
@@ -34,17 +34,13 @@ using Azure::Core::Context;
 using namespace Azure::Core::IO;
 
 // Keep reading until buffer is all fill out of the end of stream content is reached
-int64_t BodyStream::ReadToCount(
-    BodyStream& body,
-    uint8_t* buffer,
-    int64_t count,
-    Context const& context)
+int64_t BodyStream::ReadToCount(uint8_t* buffer, int64_t count, Context const& context)
 {
   int64_t totalRead = 0;
 
   for (;;)
   {
-    int64_t readBytes = body.Read(buffer + totalRead, count - totalRead, context);
+    int64_t readBytes = this->Read(buffer + totalRead, count - totalRead, context);
     totalRead += readBytes;
     // Reach all of buffer size
     if (totalRead == count || readBytes == 0)
@@ -54,7 +50,7 @@ int64_t BodyStream::ReadToCount(
   }
 }
 
-std::vector<uint8_t> BodyStream::ReadToEnd(BodyStream& body, Context const& context)
+std::vector<uint8_t> BodyStream::ReadToEnd(Context const& context)
 {
   constexpr int64_t chunkSize = 1024 * 8;
   auto buffer = std::vector<uint8_t>();
@@ -63,7 +59,7 @@ std::vector<uint8_t> BodyStream::ReadToEnd(BodyStream& body, Context const& cont
   {
     buffer.resize((static_cast<decltype(buffer)::size_type>(chunkNumber) + 1) * chunkSize);
     int64_t readBytes
-        = ReadToCount(body, buffer.data() + (chunkNumber * chunkSize), chunkSize, context);
+        = this->ReadToCount(buffer.data() + (chunkNumber * chunkSize), chunkSize, context);
 
     if (readBytes < chunkSize)
     {